Remove AD piracy functions by moving to SciMLBase
This also bumps to v1.9 because of SciMLBase, and thus drops the extra Requires.jl stuff
